--- a/src/bitmask.rs
+++ b/src/bitmask.rs
@@ -11,7 +11,7 @@ const BITMASK_NAME: &str = "bitmask.dat";
 
 #[derive(Debug)]
 pub struct Bitmask {
-    mmap: MmapBitSlice,
+    bitslice: MmapBitSlice,
     path: PathBuf,
 }
 
@@ -28,19 +28,22 @@ impl Bitmask {
 
         // length in bytes
         let length = bits / 8;
-
         length
     }
 
     /// Create a bitmask for one page
-    pub fn new(dir: PathBuf, page_size: usize) -> Self {
+    pub fn with_capacity(dir: PathBuf, page_size: usize) -> Self {
         let length = Self::length_for_page(page_size);
+
         let path = Self::bitmask_path(dir);
         create_and_ensure_length(&path, length).unwrap();
         let mmap = open_write_mmap(&path, AdviceSetting::from(Advice::Normal), false).unwrap();
         let mmap_bitslice = MmapBitSlice::from(mmap, 0);
+
+        assert_eq!(mmap_bitslice.len(), length * 8, "Bitmask length mismatch");
+
         Self {
-            mmap: mmap_bitslice,
+            bitslice: mmap_bitslice,
             path,
         }
     }
@@ -50,7 +53,7 @@ impl Bitmask {
         let mmap = open_write_mmap(&path, AdviceSetting::from(Advice::Normal), false).unwrap();
         let mmap_bitslice = MmapBitSlice::from(mmap, 0);
         Self {
-            mmap: mmap_bitslice,
+            bitslice: mmap_bitslice,
             path,
         }
     }
@@ -60,7 +63,7 @@ impl Bitmask {
     }
 
     pub fn infer_max_page_id(&self, page_size: usize) -> usize {
-        let length = self.mmap.len();
+        let length = self.bitslice.len();
         let bits = length * 8;
         let covered_bytes = bits * BLOCK_SIZE_BYTES;
         covered_bytes / page_size
@@ -71,14 +74,14 @@ impl Bitmask {
         let extra_length = Self::length_for_page(page_size);
 
         // flush outstanding changes
-        self.mmap.flusher()().unwrap();
+        self.bitslice.flusher()().unwrap();
 
         // reopen the file with a larger size
-        let new_length = self.mmap.len() + extra_length;
+        let new_length = (self.bitslice.len() / 8) + extra_length;
         create_and_ensure_length(&self.path, new_length).unwrap();
         let mmap = open_write_mmap(&self.path, AdviceSetting::from(Advice::Normal), false).unwrap();
 
-        self.mmap = MmapBitSlice::from(mmap, 0);
+        self.bitslice = MmapBitSlice::from(mmap, 0);
     }
 
     fn range_of_page(page_id: PageId, page_size: usize) -> Range<usize> {
@@ -95,19 +98,36 @@ impl Bitmask {
     /// The amount of blocks that have never been used in the page.
     pub(crate) fn free_blocks_for_page(&self, page_id: PageId, page_size: usize) -> usize {
         let range_of_page = Self::range_of_page(page_id, page_size);
-        self.mmap[range_of_page].trailing_zeros()
+        self.bitslice[range_of_page].trailing_zeros()
     }
 
     /// The amount of blocks that are available for reuse in the page.
     pub(crate) fn fragmented_blocks_for_page(&self, page_id: PageId, page_size: usize) -> usize {
         let range_of_page = Self::range_of_page(page_id, page_size);
-        let bitslice = &self.mmap[range_of_page];
+        let bitslice = &self.bitslice[range_of_page];
 
         bitslice.count_zeros() - bitslice.trailing_zeros()
     }
 
-    pub(crate) fn find_available_blocks(&self, num_blocks: u32) -> Option<(PageId, BlockOffset)> {
-        todo!();
+    pub(crate) fn find_available_blocks(
+        &self,
+        num_blocks: u32,
+        page_size: usize,
+    ) -> Option<(PageId, BlockOffset)> {
+        let mut block_cursor = 0;
+        while (block_cursor + num_blocks as usize) < self.bitslice.len() {
+            let bitslice = &self.bitslice[block_cursor..block_cursor + num_blocks as usize];
+            if let Some(offset) = bitslice.last_one() {
+                // skip the whole part which has ones in the middle
+                block_cursor += offset + 1;
+            } else {
+                // bingo - we found a free cell of num_blocks
+                let page_id = block_cursor.div_euclid(page_size);
+                let block_offset = block_cursor.rem_euclid(page_size);
+                return Some((page_id as PageId, block_offset as BlockOffset));
+            }
+        }
+        None
     }
 
     pub(crate) fn mark_blocks(
@@ -119,8 +139,47 @@ impl Bitmask {
     ) {
         let page_start = Self::range_of_page(page_id, BLOCK_SIZE_BYTES).start;
 
-        let blocks_range = page_start + block_offset as usize..num_blocks as usize;
+        let offset = page_start + block_offset as usize;
+        let blocks_range = offset..offset + num_blocks as usize;
 
-        self.mmap[blocks_range].fill(used);
+        self.bitslice[blocks_range].fill(used);
     }
+}
+
+#[cfg(test)]
+mod tests {
+
+    #[test]
+    fn test_length_for_page() {
+        assert_eq!(super::Bitmask::length_for_page(8192), 8);
+    }
+
+    #[test]
+    fn test_find_available_blocks() {
+        let page_size = 8192;
+        let dir = tempfile::tempdir().unwrap();
+        let mut bitmask = super::Bitmask::with_capacity(dir.path().to_path_buf(), page_size);
+        // 1..=10
+        bitmask.mark_blocks(0, 1, 9, true);
+
+        // 15..=20
+        bitmask.mark_blocks(0, 15, 5, true);
+
+        let (page_id, block_offset) = bitmask.find_available_blocks(1, page_size).unwrap();
+        assert_eq!(page_id, 0);
+        assert_eq!(block_offset, 0);
+
+        let (page_id, block_offset) = bitmask.find_available_blocks(2, page_size).unwrap();
+        assert_eq!(page_id, 0);
+        assert_eq!(block_offset, 10);
+
+        let (page_id, block_offset) = bitmask.find_available_blocks(6, page_size).unwrap();
+        assert_eq!(page_id, 0);
+        assert_eq!(block_offset, 20);
+
+        let found_large = bitmask.find_available_blocks(100, page_size);
+        assert_eq!(found_large, None);
+    }
+
+    // TODO: proptest!!! (for find_available blocks)
 }