--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod fixtures;
+mod page;
 mod page_tracker;
 pub mod payload;
 mod payload_storage;
-mod slotted_page;
 mod stats;
 mod utils_copied;