Coauthored by @agourlay. Most of the work was done on pair programming sessions.

Updates everything to match the new bitmask design with headerless pages.

## TODO
- [x] introduce gaps structure to find empty spots quicker.