@coszio 2024-10-17
Just pushed the changes to add `Gaps` structure to each region. (Coauthored by @agourlay again).

However, there is a test that is now failing because we changed the logic to creating new pages, can add an extra page. 
We need to optimize this part to fix the test.

Additionally, although the `Gaps` structures improve performance, it is still quite a lot slower than previous implementation. Let's discuss ways to improve 2 things:
- replace `BitSlice::last_one` with something faster.
- put the `Gaps` in a different structure so we don't have to do a linear scan to find a fitting gap.