@coszio 2024-10-18
Optimization update for:
- `find_available_blocks`
- `calculate_gaps`

These two were slowest functions when profiling, and have been optimized to run almost 10x faster.

They now process the bitslices in chunks of 64 bits to take advantage of bitwise operators over `usize`.

## Flamegraphs

Before optimizations
![imagen](https://github.com/user-attachments/assets/90b5c93c-ea3c-4f5d-9f51-af9a50bbb8e8)


After optimizing `calculate_gaps`
![imagen](https://github.com/user-attachments/assets/7c14566c-8a9f-490f-abe7-704815f9f004)


After optimizing `find_available_blocks`
<img width="898" alt="imagen" src="https://github.com/user-attachments/assets/1c3fb9f7-e1b7-4655-ab23-0313cb9c7713">

## Benchmarks
Before opts:

```**read_heavy** with prefill_fraction 0.95
PayloadStorage:
1572864 operations across 1 thread(s) in 4.020268542s; time/op = 2.556µs
RocksDB:
1572864 operations across 1 thread(s) in 937.348584ms; time/op = 595ns

**insert_heavy** with prefill_fraction 0.0
PayloadStorage:
1572864 operations across 1 thread(s) in 82.551443375s; time/op = 52.484µs
RocksDB:
1572864 operations across 1 thread(s) in 3.69300475s; time/op = 2.347µs

**update_heavy** with prefill_fraction 0.5
PayloadStorage:
1572864 operations across 1 thread(s) in 30.615338042s; time/op = 19.464µs
RocksDB:
1572864 operations across 1 thread(s) in 2.141997875s; time/op = 1.361µs
```


After opts:
```
**read_heavy** with prefill_fraction 0.95
PayloadStorage:
1572864 operations across 1 thread(s) in 1.23288875s; time/op = 783ns
RocksDB:
1572864 operations across 1 thread(s) in 847.860959ms; time/op = 539ns

**insert_heavy** with prefill_fraction 0.0
PayloadStorage:
1572864 operations across 1 thread(s) in 6.560875792s; time/op = 4.171µs
RocksDB:
1572864 operations across 1 thread(s) in 3.393870416s; time/op = 2.157µs

**update_heavy** with prefill_fraction 0.5
PayloadStorage:
1572864 operations across 1 thread(s) in 4.655892083s; time/op = 2.96µs
RocksDB:
1572864 operations across 1 thread(s) in 2.005825375s; time/op = 1.275µs
```

**TLDR;** It is still slower than RocksDB, but not on a different league.

I still believe we can make this faster by putting some more brainpower into this.