@timvisee 2024-10-21
> They now process the bitslices in chunks of 64 bits to take advantage of bitwise operators over usize.

I assume you're talking about using the `usize` store type in bit slices. Could you check what performance is like with these two configurations as well:
- [`Msb0`](https://docs.rs/bitvec/latest/bitvec/order/struct.Msb0.html) bit order and `u8` bit store
- [`Lsb0`](https://docs.rs/bitvec/latest/bitvec/order/struct.Lsb0.html) bit order and `usize` bit store

The above two allow portability across different architectures, but may have a slight performance impact.